BJ Refactor: Move DealerHand class to DealerHand.cs
The DealerHand class has been relocated from Hand.cs to a new file named DealerHand.cs. This includes moving the class definition and its methods (FlipSecondCard and FlipCard) to the new file. The DealerHand class is now part of the Demo.Demos.BJ namespace. The functionality of the methods remains unchanged.